Run asset precompilation in rails production mode
As part of our goal to have all our applications use `buildProject` in
their Jenkinsfile we realised that whitehall ran its asset precompilation
in RAILS_ENV=production.  Rather than allow an overridePrecompileAssets
variable, I think the better option is to make this behaviour the default.

The original [commit to whitehall][1] that introduces this has a good
rationale for this.  When we deploy we build the assets in production mode
which has differences in behaviour to development or test (for example
it does digests) and so a better test is to run the asset precompilation
in production mode to catch any problems that may occur before we try to
deploy.

We also add a bunch of asset-y env vars to the rake task. These are set
by whitehall, and required by Plek to build absolute urls. In development
and test mode they're either not used or have hardcoded default values
which is not the case in production mode. In deployed environments the
values come from govuk-puppet and `govuk_setenv`, but that's not in use
on our jenkins infrastructure.

[1]: https://github.com/alphagov/whitehall/commit/231ff4996305788a466f009d5e46625376f8b30a