Remove relationship to pass tests
This complained about a relationship error, despite the class seemingly
being included when testing in Vagrant:

```
govuk::node::s_calculators_frontend should compile
     Failure/Error: expect { subject.call }.not_to raise_error
       expected no Exception, got #<Puppet::Error: Could not find resource 'Class[Govuk::Apps::Calculators]' for relationship from 'Govuk_mount[/data/vhost]' on node calculators-frontend-1.example.com> with backtrace:
```

This appears to have been added as a temporary measure but never
removed (ref: e16e1af91eaefa48437a365a511fba99025745fc).

The conditional was then removed (ref: e84432d5d28dbae5708e04ae89c4e9c79c3cf8bf)

This is the *only* app that requires a mount before the app is deployed,
and I cannot see anything special about this app as to why it would be
the exception in this case.

It bothers me that I unable to find out why this started to error, which
makes me think that there is a relationship issue that I am missing, but
this should be removed to be consistent.