Make enable_authenticating_proxy parameter clearer
When checking this I found it extremely confusing. We inherit a class,
and set a parameter in the hieradata for a different node type.

It would be clearer to call the class so we can understand what is going
on here.

This change will make it more obvious that authenticating proxy is
disabled on s_cache, but enabled on s_draft_cache.