Add designprinciples back to frontend classes
Our previous PR (#6781) to ensure design principles is removed from our
deployed servers was too zealous.  Removing the designprinciples entry
from the frontend class definition means the `ensure => 'absent'` in
designprinciples.pp is meaningless because puppet won't use that file
anymore on the machines it's deployed to.  Adding it back means it will
understand that it still has to deal with designprinciples.pp and this
will actually remove the designprinciples app from those machines.