Use docker pip dependency rather than docker-py
docker-py has been renamed to just docker. The docker-py
dependency had compatibility issues with docker-compose so this now uses
docker as suggested by docker-compose:
https://github.com/docker/compose/issues/4569

We've had to use the exec method to install the dependency as we hit a
duplicate namespace issue if we install it via package. This issue is
outlined here: https://tickets.puppetlabs.com/browse/PUP-1073 and seems
to be fixed in Puppet 4.