Introduce statsd repeating into graphite.cluster
We have identified a problem with our statsd configuration that affects
statistics that can be sent with the same key from multiple hosts.
Graphite will only store the last value sent within a window (this is
currently configured to 5 seconds). If two hosts send a statistic with
the same key within this window only the last one will be recorded and
may mean for some rapidly changing statistics our data represents a 1/2
or 1/3 of the actual value.

This PR introduces a plugin to statsd - statsd_smart_repeater - which is
used to send the aggregated statistics of a host to another statsd
instance which then sends this data to graphite. This second statsd
instance will do the aggregation between hosts.

The scope of this PR is to introduce this functionality in addition to
current behaviour (with all the stats behind a namespace of
smart_repeater) to allow comparison with existing behaviour.

The script for statsd_smart_repeater is from
https://github.com/lucidsoftware/statsd-smart-repeater this was used
instead of the repeater packaged with statsd as that does not support
aggregating stats. A slightly more popular alternative seems to be
https://github.com/wanelo/gossip_girl however this was rejected due to
it not supporting TCP (and it's a little less feature rich e.g.
namespaces).