[AWS] Fix reboot check
Currently the unattended reboot check is run from monitoring-1, and
connects to puppetdb over http. It queries whether a class includes
either `govuk_safe_to_reboot::yes` or `govuk_safe_to_reboot::no`, and
reports to Icinga which machines need to be rebooted.

We do not allow connections over http to puppetdb in AWS, so we must
move the check to the puppetmaster itself, and set up an alias so that
the script `govuk_node_list` queries localhost using the correct host
header, rather than going back out and in through the load balancer.