Inline setEnvar call into initializeParameters
After making `initializeParameters` part of the standard `buildProject`
again we started getting reports of failed builds.  The builds would fail
with the following error:

    java.lang.NoSuchMethodError: No such DSL method 'setEnvar'

and the following relevant stack trace entries:

    at Script1.initializeParameters(Script1.groovy:395)
    at Script1.buildProject(Script1.groovy:130)

This points us at the invocation of `initializeParameters` we added to
the start of `buildProject`.  It's entirely unclear why this call to
`initializeParameters` can't find `setEnvar`, it's presumably a scope
issue, but I can't work it out.  More curious is that the same builds
would continue to fail if build in response to new commits, but would
sometimes start to pass if built manually, or replayed.  Once the build
passes once (or fails for a real reason - e.g. it gets past the call to
`initializeParameters`) it would continue to run properly.

The workaround seems to be inlining the code form `setEnvar` into
`initializeParameters` directly to avoid the scoping issue.  An
alternative would be to stop calling `initializeParameters` but we know
that first time builds of some projects (publishing-api in particular)
will fail if we don't call it because of a Jenkins bug (JENKINS-40574)
meaning env vars aren't properly set to their default values from build
parameters for 1st time builds in a pipeline.