Upgrade UNKNOWN to CRITICAL
This alert appears to be raised when there is a problem with the Puppet
summary file, which more often than not means that it has not been
written correctly. As an example:

```
---
  version:
    config:
    puppet: "3.8.5"
  time:
    last_run: 1511542072
```

We think this is when Puppet cannot receive the catalogue, which means
it doesn't even write any failures in the state.

UNKNOWN alerts tend to go unnoticed by 2ndline, and being unable to
retrieve the catalogue is usually related to a relationship or duplicate
resource error. These should be fixed as priority, so upgrade the alert
to CRITICAL, rather than WARNING.