Grafana dashboard for Email Alert API Metrics
This dashboard is to assist the Email team in understanding key metrics
that affect performance of the Email Alert API as it grows in scope.
This iteration of this dashboard only contains statistics that are
related with the Notify integration, however the intention is that this
will continue into more data that cover wider concerns.

This dashboard was built through the Grafana UI and this file was
created by exporting that data.