Collect sidekiq logs for apps that log
I would like to collect logs for Publisher's workers, but feel it might be
useful to collect them for everything (and hopefully get closer to a consistent
logging experience across our apps).

Some apps log to sidekiq.json.log and some to sidekiq.log but it all seems to
be in json format so I think we should be OK.

If there are multiple files that the path matches then Filebeat will
[set up a harvester for each](https://www.elastic.co/guide/en/beats/filebeat/master/configuration-filebeat-options.html#prospector-paths)

I think it follows that if it doesn't find anything the, that's OK too.