Add the build token root plugin
https://plugins.jenkins.io/build-token-root

This plugin describes the problem it is solving better than I can -

"The use case is that Jenkins is secured so that anonymous users lack
overall read permissions. Say you want to triggers builds of certain
jobs from a script. You can pick a sufficiently authenticated user and
use that person’s API token to POST to job/NAME/build. But this grants
that person’s full permissions to anyone who can see the script, which
is hazardous.

The usual workaround for this issue is to define a build authorization
token in job configuration, and have the script ping
job/NAME/build?token=SECRET. Unfortunately Jenkins checks URIs
hierarchically and just getting as far as job/NAME/ requires
authentication."

By adding this plugin, we dont have to manage auth tokens for users
which requires a bootstrapping process because jenkins users cannot be
fully managed. There is a small risk that frontend servers could trigger
deploys on backend servers, but they cannot pass parameters to these
builds