Enable email-alert-api-public proxy
This adds a hostname, email-alert-api-public, that is exposed
publicly which is used as a means for external services to route
requests in to email-alert-api.

We decided to go with the unconventional name of email-alert-api-public
as it wasn't clear about all the usages of this hostname might be and
this seemed to be the clearest about the intent of it.

It proxies the endpoint status-updates and will accept any subroutes
that start with that segment. Any other requests will receive a 404.

This is currently only enabled on integration until we have completed
authentication work.