Output hostnames based on instance-ids
The `govuk_node_list` tool is very useful for working inside the
environment. In AWS we can gather node classes by tags, but if we want
to work with PuppetDB, then we need to convert InstanceIDs to
hostnames.

We moved to using InstanceIDs so we did not run into conditions of
repeated hostnames. In a DHCP environment, hostnames are reused, and so
this was potentially conflicting in PuppetDB. With an InstanceID, we can
guarantee a unique identifier.

We have checks for unattended reboots, that queries PuppetDB for
specific classes. In the traditional way, this was outputting a list of
InstanceIDs. While these IDs may be queried, it doesn't give you a way
to immediately connect to the relevant hostname.

I've chosen to put the logic into `govuk_node_list` as this function is
to list hostnames. The conversation does a look-up based upon the
InstanceID, and returns the hostname.

We could potentially use this with our fabric scripts as well, which
would mean less deviation from the way we currently support the system.