Add healthcheck path to default vhost
In AWS, we are looking at using the L7 Application Load Balancers. To
ensure nodes are in service on some instances before we amend Puppet to
go straight to the correct port as defined by host header, we are adding
in a default healthcheck path to allow the node to be in service on the
load balancer.