Strip slashes from docker image tags
We pass branch names in to the Jenkins `buildDockerImage` function to
be used as a tag for the Docker image. However, a branch name might
contain a slash (eg `feature/awesome-new-feature`), which is not valid
in the Docker tag name.

In this case, the build will fail.

This change adds a method for replacing these slashes with underscores,
which are allowed in the tag name.