Configure a Slack channel for the CDN dictionary Jenkins job
Add a Slack channel to the `publishers` section of the CDN dictionaries
Jenkins job to fix an error in the Jenkins job builder.

We don't think this channel is actually used by the Jenkins job since it
always falls back to the default channel, but apparently it needs to be
configured anyway.