Use docker system prune to clean up docker artefacts
docker system prune [1] is a more modern command that we can use to
clean up the various things that are left behind by docker builds. It
removes all things that are older than 24 hours and not in use - so this
may cause some additional building of containers but I'm expecting
effect to be low.

Running this on ci-agent-4 reclaimed 75.75GB space

This uses docker volume prune to remove the images as they are not
removed with `docker system prune` after version 17.06 and do not
support the filter command.

[1]: https://docs.docker.com/engine/reference/commandline/system_prune/