Improve the signon queue size check
Remove the summarize, as this duplicates what the check_graphite
script does. Customise the from and dropfirst arguments to get data
from the last 24 hours, but focus in on the last minute for the
alert.

This change is a result of changing the statsd configuration to delete
gauges. As statsd isn't sending old values anymore, this check needed
adjusting to not fail on missing data.