Improve the rummager queue size check
Remove the movingAverage, as this duplicates what the check_graphite
script does. Increase the from argument to 24 hours to account for
missing data, and use dropfirst to focus on the most recent ~3 minutes
worth of data.

This change is a result of changing the statsd configuration to delete
gauges. As statsd isn't sending old values anymore, this check needed
adjusting to not fail on missing data.