Don't run Publishing end-to-end tests for schema builds
When govuk-content-schemas builds it triggers tests for many
applications, and these applications then kick off end-to-end tests.
Which can lead to a huge amount of builds. These end-to-end test builds
don't actually offer any value since we don't offer the ability to build
against a particular content-schema commit.

A future iteration could be that govuk-content-schemas kicks off it's
own build of the end-to-end tests once we support specifying the build.