Add extra CI agents to run end-to-end tests
This adds 3 additional CI agents to both Carrenza and AWS environments.
It adds a label for running the [Publishing end-to-end tests][1] on just
particular CI agents. To split the tasks it uses 4 boxes for end-to-end
testing and 4 for all other jobs.

Hopefully this should isolate end-to-end tests from affecting other
builds and with e2e performance improvements we might be able to reduce
this or redistribute the load.

At the moment end-to-end tests cannot run in parallel so this sets the
executors to 1 so that we don't have a bunch of locked builds.

[1]: https://github.com/alphagov/publishing-e2e-tests