Explicitly prevent S3's Content-Type header reaching the client
This hopefully fixes
https://github.com/alphagov/asset-manager/issues/321.

By explicitly telling nginx not to forward certain headers from S3 to
the client, we should prevent a potential race condition or undefined
ordering behaviour in nginx where the empty Content-Type field in S3
takes precedence over the one set by Rails.

This causes the issue as all(?) assets in S3 have a blank Content-Type
field, which S3 returns as an empty field.