Add Licensify Plek override for AWS
We will not be hosting Licensify in AWS, so we need to redirect
Licensify traffic over to the current environment.

Currently the Frontend app uses Plek to find the right host, so we need
to override this and provide the original service domain.

We set the host currently in the /etc/hosts file using an external IP,
but in this instance we are going to create a DNS record so we do not
have to do this.