Remove Oracle8 from govuk_ci agents
There is an issue installing JDK8 because Oracle have removed the
download location and doesn't work.

We only include this for Licensing, which we don't test on this CI
environment anyway.

I've put it in a conditional to make it clear incase we ever come back
to it and wonder (or want to try enabling on a single agent).