Include JDK if Elasticsearch is not included
ci-agent-6 specifically failed to join the Jenkins master as an agent
because a JDK was not installed. This was because it had been explicitly
excluded from installing Elasticsearch (probably from when we were using
it to test Xenial).

If we're not including Elasticsearch then we should include a JDK
instead.