Update default resolvers
We defaulted to using Google resolvers, but recently they have
implemented some rate-limiting which was noted when we ran the "Validate
Published DNS" job.

Carrenza provide resolvers, so use these instead to ensure we do not
suffer from timeouts when making lookups.

In AWS, we use the VPC for resolving and so this is not included.