Set resolvers for UKCloud servers to Google
I updated the resolvers in 15f15ccf7d8002bd61e58667d8077d143f51b301 to
the Carrenza resolvers.

Carrenza only allow Carrenza IPs to access their resolvers, and
Licensing is a special snowflake that lives in UKCloud.

Set the resolvers back to the Google ones for UKCloud in Staging and
Production. Integration is not affected as that is in Carrenza.