Increase maximum number of postgres connections
We're looking at increasing the number of workers in one of our
apps from 30 to 90 in staging and production. This will add
another 60 connections to each environment, of which they both
have about 30 spare connections available. Looking at the postgres
machines they both have lots of spare memory capacity available to
cope with this level of increase.