Remove Sendfile-related config from asset-manager Nginx vhost
Since alphagov/asset-manager#328 the Rails app only ever proxies asset
requests to S3; it no longer instructs Nginx to serve asset files from
NFS.

The functionality for proxying to S3 does not rely on setting the
X-Sendfile-Type or X-Accel-Mapping request headers - these are only
needed when using the Rack::Sendfile middleware built in to Rails which
we are no longer using. So we can remove the location for /media/ &
/government/uploads/ and instead rely on the default location for /
which is provided by proxy-vhost.conf [1].

Similarly the Rails app will no longer ever use the /raw/ prefix in the
X-Accel-Redirect response header value and so we can also remove the
corresponding location block.

In this commit I've also added an explanatory comment to the
/cloud-storage-proxy/ location block to replace the comment which
previously existed against the /media/ & /government/uploads/ location
block.

[1]:
https://github.com/alphagov/govuk-puppet/blob/1485ebb5a8ce228f6d4e60e8329d05e6c95a0832/modules/nginx/templates/proxy-vhost.conf#L139-L142