Don't send GOVUK-Asset-Manager-File-Store response header
Now that asset requests are only ever proxied to S3, there's no point in
this header which was used to discriminate between responses which were
served from NFS vs proxied to S3.

Since this header will no longer be sent, there's no point in attempting
to include it in the log output.