Adjust the support queue size check for a change to statsd
Increase the from argument to 24 hours to account for missing data (as
the Sidekiq statsd middleware only reports metrics when there is
activity), and use dropfirst to focus on the most recent ~3 minutes
worth of data.

This change is a result of changing the statsd configuration to delete
gauges. As statsd isn't sending old values anymore, this check needed
adjusting to not fail on missing data.