Add postgres client to jenkins boxes
We're running a jenkins job that checks out an app and runs a script from it.

Technically the app doesn't need to be using a database right now at all, but we
want it to use one in the future, so adding the pg library is a simpler workaround
than changing the rails configuration.

Normally, when jenkins runs a script thats package up with an app, the jenkins
job ssh's into the box that runs it the app and runs the script from there. In
this case the app runs on heroku so it's not as straightforward to run it
remotely.

The alternative is to just remove the search benchmark jobs altogether.