Increase the window size for fetching Signon sidekiq data
This is showing as Unknown on Staging, which is a false positive, as
the metrics are still valid, it's just that nothing has happened, so
the Sidekiq middleware hasn't sent any data.

A very large window isn't ideal, as this could hide a lack of data, so
just increase it a little bit, to hopefully reduce the false positives
on Staging.