Alternative way to test pre-migration
This is a really horrible way to test the stack pre-migration, but I
can't think of another way around it.

Both sets of Smokey tests (in Icinga and Jenkins) rely on running
against backend publishing apps. I tried setting the
"SPOOF_TARGET_DOMAIN" env var, but this did not take into account the
backend apps, which means we can't fully test the stack.

Setting these hosts will also enable us to log into backend apps, as
otherwise they try to authenticate with Signon in the current
Integration, which means we are unable to login to them.

The IPs are liable to change because of the way ELBs are dynamic, but we
can always regenerate them using the tool created in [govuk-aws](https://github.com/alphagov/govuk-aws/blob/master/tools/migration-hosts.sh).

Most are aliases to the same place, so can use the same IP.