Change sidekiq-monitoring to listen on port 3211
When we removed the `localhost` part of the listen definition in
60a89ae9a7a30406225a3ff37cf17928941fd427 the instructions for tunnelling
into backend machines to view the sidekiq monitoring apps stopped
working.  I think because we try to tunnel into localhost:80 and the
sidekiq monitoring app doesn't respond to localhost:80, something else
does.  If we change the port to something else then it starts working
again, and also means it'll work on AWS because lots of things already
listen on port 80 in AWS so this had no chance of working.

We pick 3211 as it's the next available number in the development-vm
Procfile, which in the absence of anything else more official, acts as
the documentation for which ports are in use by our apps.