Add support for GOVUK_APP_DOMAIN_INTERNAL
In
https://github.com/alphagov/smokey/commit/8061d3783c59f1a2c83a4aafb4da2553aa3290fc,
support was added for the GOVUK_APP_DOMAIN_INTERNAL environment
variable. This allows us to specify internal domains in our Smokey tests
as there are some services, such as Bouncer, which we do not expose
publically to the internet, so have to route internally.

This was traditionally solved by having hardcoded hosts entry in
/etc/hosts, but is not ideal to do this in AWS.

This commit also sorts the list of PLEK_SERVICE environment variables.