Make the email-alert-check jobs resilient to GitHub being down
https://trello.com/c/0nNVddWb/408-%EF%B8%8F-dont-call-people-when-github-is-down

Previously, the build failed if the clone from GitHub failed. We'd tried
to use the Text Finder plugin to change the job status from 'failed' to
'unstable' so that Failure_Passive_Check wouldn't trigger post-build,
but this approach doesn't work.

The problem is the Text Finder plugin can only downgrade build statuses,
e.g. success -> unstable, it can't go fail -> unstable. See:
https://issues.jenkins-ci.org/browse/JENKINS-19641

This change makes it so we guard against the clone to GitHub failing by
cloning to a 'tmp' directory, which only replaces the existing
email-alert-monitoring directory if the clone succeeds. Otherwise, we
echo from this command which causes execution to continue. This
has the effect of re-using the previously cloned directory from the last
time the job ran.

If the clone fails and the email check succeeds, we mark the build as
unstable by matching the 'Cloning failed' text in the build output. This
shouldn't alert in PagerDuty, but will add visibility that something
isn't quite right.

While working on this, I also looked at the Conditional Build Plugin and
Groovy Script Plugin, but I didn't manage to get either of these plugins
to trigger when the clone fails using Jenkins' default SCM settings for
the repository.