Set everything as internal in AWS
Set "GOVUK_APP_DOMAIN" as internal by default. Ensure that each
application sets a Plek override for itself so that if it ever
references itself, it redirects the user to the correct location.

This hopefully solves a bunch of problems we have come across when
trying to correctly route traffic internally. We had an issue where
applications would misuse Plek to find out what their URL is to redirect
users.