Rework Smokey environment variables
When trying to understand how we deploy Smokey, I was a little bemused
by the confusing way how we set environment variables in two different
ways, even though we generally run the app using `govuk_setenv`.

It would be more straight forward to understand using Smokey both in
Icinga and Jenkins if we were to set the environment variables in the
same way they are set for all other applications, and in a single place
that both the tasks can consume.

This also sets GOVUK_APP_DOMAIN specifically for Smokey in AWS. Smokey
relies on correctly authing with Signon for many of the publisher
application tests, and if we make a request using an internal domain,
then it will send the callback to the wrong place. Ensuring the
GOVUK_APP_DOMAIN environment variable is overwritten resolves this
issue.

This relies on the change in Smokey itself: https://github.com/alphagov/smokey/commit/964f88bad7fe0134645396dc9fc7275e3c2fb340