Add Plek Signon override
We need to set a specific override for Plek as otherwise users are
redirected back to the internal domain.

Applications use Plek to redirect users to Signon for auth, and without
this override being set Plek returns an incorrect value for an end user:

```
irb(main):001:0> Plek.find("signon")
=> "https://signon.integration.govuk-internal.digital"
```

Signon redirects users back to the application using what is set with
the `redirect_url` parameter in Signon for whichever application, so we
only need to set application specific behaviour.