Change docker prune approach
The approach of doing a system prune -a every 2 hours seemed a bit too
aggressive as it was causing a build failure if a docker container was
being started at the point this was run.

This goes for the less aggressive just dangling option to run every 2
hours and then at 4am each day to run a more aggressive prune to remove
the non-dangling images too (hopefully it's less likely someone will be
running a build then).

Hopefully net effects of this should be that the disk space is kept at
bay whilst build failures are less and build times lower.