Fix date range for 'detailed' search index size check
When checking that the `detailed` search index hasn't changed
significantly in size, compare the current size with the data from 7
days ago rather than 3.

This is important because integration is switched off over the weekend.
Comparing a weekday with a weekend will cause spurious alerts because
graphite will try to compare a full search index with a missing one.

The other indices were already using 7 days for the comparison. This
just makes the `detailed` index consistent with the others.