Display sidekiq graphs & stats on Asset Manager deployment dashboard
I suspect these were not added originally, because Asset Manager used to
use Delayed Job rather than Sidekiq.

I think these will be a useful addition, particularly now that we have
two different workers - one for virus scanning and the other for
uploading to S3 - both of which are integral to the app working
correctly.